backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2023.10.0
huggingface-hub==0.32.2
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
packaging==25.0
pfzy==0.3.4
prompt_toolkit==3.0.51
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
django-cors-headers==4.3.1
datasets==2.14.6
accelerate==0.24.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
